moved all components to /components
